# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OwnerCostResponse": {
        "additionalProperties": false,
        "description": "``GET /api/v1/cost/per-owner`` response.\n\n``from_ts`` / ``to_ts`` echo the request filters (ISO 8601) so\na client can confirm the window the totals were computed for\nwithout re-parsing the query string.",
        "properties": {
          "from_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Ts"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnerCostSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "to_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Ts"
          }
        },
        "required": [
          "items"
        ],
        "title": "OwnerCostResponse",
        "type": "object"
      },
      "OwnerCostSummary": {
        "additionalProperties": false,
        "description": "One row in :class:`OwnerCostResponse.items`.\n\n``owner`` is the API key / dashboard label; ``None`` when the\nsession was submitted before Plan 7 D7.26 added the owner\ncolumn (legacy rows). The dashboard renders ``None`` as\n``\"\u2014\"`` to keep the table column aligned.",
        "properties": {
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "session_count": {
            "default": 0,
            "title": "Session Count",
            "type": "integer"
          },
          "total_cost_usd": {
            "default": 0.0,
            "title": "Total Cost Usd",
            "type": "number"
          }
        },
        "title": "OwnerCostSummary",
        "type": "object"
      },
...
      "SessionCostBreakdown": {
        "additionalProperties": false,
        "description": "One row in :class:`SessionCostListResponse.items`.\n\nCompact shape \u2014 only the fields the cost breakdown table needs;\nno spec dump or frames. ``ended_at`` is included so the dashboard\ncan render the elapsed wall-clock alongside cost for incomplete\nruns that already accumulated tokens.",
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "total_cost_usd": {
            "default": 0.0,
            "title": "Total Cost Usd",
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "submitted_at"
        ],
        "title": "SessionCostBreakdown",
        "type": "object"
      },
      "SessionCostListResponse": {
        "additionalProperties": false,
        "description": "``GET /api/v1/cost/per-session`` response.\n\nCursor pagination shape mirrors :class:`SessionListResponse` so\na future ``after=...`` upgrade is non-breaking; the Task-23 MVP\nalways returns ``next_cursor=None``.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionCostBreakdown"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "SessionCostListResponse",
        "type": "object"
      },
...
      "UserCostSummary": {
        "additionalProperties": false,
        "description": "``GET /api/v1/cost/summary`` response.\n\n``team_total_cost_usd`` is admin-only \u2014 non-admin callers see\n``None``. The dashboard renders it as the \"team\" stat next to\nthe per-user stat so admins can compare their personal share\nagainst the team total at a glance.",
        "properties": {
          "from_ts": {
            "title": "From Ts",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "session_count": {
            "default": 0,
            "title": "Session Count",
            "type": "integer"
          },
          "team_total_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Total Cost Usd"
          },
          "total_cost_usd": {
            "default": 0.0,
            "title": "Total Cost Usd",
            "type": "number"
          },
          "user": {
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "role",
          "period",
          "from_ts"
        ],
        "title": "UserCostSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "gg-relay",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cost/export.csv": {
      "get": {
        "description": "CSV export of the per-owner aggregate \u2014 admin only.\n\nPlan 8 D8.30 / Task 23. Writes a single ``audit_log`` row with\n``action='cost_export'`` so every download is traceable. The\n``target_id`` carries the requested time window so a follow-up\naudit query can correlate downloads with billing periods.\n\nCapped at 1000 rows \u2014 single-team scale fits easily; larger\ndeployments should drive ad-hoc Postgres queries.",
        "operationId": "cost_export_csv_api_v1_cost_export_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cost Export Csv",
        "tags": [
          "cost"
        ]
      }
    },
    "/api/v1/cost/per-owner": {
      "get": {
        "description": "Aggregate cost per owner (admin sees all; submitter forced self).\n\nPlan 8 D8.30 / Task 23. ``order_by`` controls the secondary\nordering of the response items:\n\n  * ``cost``     \u2014 DESC by ``total_cost_usd`` (default; top\n                   spenders first).\n  * ``sessions`` \u2014 DESC by ``session_count`` (heaviest users\n                   by activity).\n  * ``owner``    \u2014 ASC alphabetical (stable diff-friendly\n                   order for CSV consumers).\n\nThe optional ``owner`` filter narrows the GROUP BY result to a\nsingle row; submitters cannot pass an ``owner`` other than\ntheir own label (403 ``forbidden_cost_view``).",
        "operationId": "cost_per_owner_api_v1_cost_per_owner_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "from_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "default": "cost",
              "pattern": "^(cost|sessions|owner)$",
              "title": "Order By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerCostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cost Per Owner",
        "tags": [
          "cost"
        ]
      }
    },
    "/api/v1/cost/per-session": {
      "get": {
        "description": "List sessions with per-row cost (newest first).\n\nPlan 8 D8.30 / Task 23. Mirrors :func:`cost_per_owner`'s RBAC:\nsubmitters are forced to ``owner=<self-label>``. ``next_cursor``\nis always ``None`` in the MVP \u2014 operators with large result\nsets should use the CSV export endpoint instead.",
        "operationId": "cost_per_session_api_v1_cost_per_session_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "from_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCostListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cost Per Session",
        "tags": [
          "cost"
        ]
      }
    },
    "/api/v1/cost/summary": {
      "get": {
        "description": "User-centric summary with 30s TTL cache.\n\nPlan 8 D8.30 / Task 23. Cache key ``(label, period, is_admin)``\n\u2014 admin role affects whether ``team_total_cost_usd`` is\npopulated so it MUST be in the key. A second call within the\nTTL hits the cache without touching the store at all.\n\nAdmin callers get the team-wide total computed from a single\nextra :meth:`aggregate_cost_by_owner` call (limit=1000); for\nlarger team sizes consider denormalising into a daily snapshot\ntable (future plan).",
        "operationId": "cost_summary_api_v1_cost_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "this_month",
              "pattern": "^(today|this_month|last_30d)$",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCostSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cost Summary",
        "tags": [
          "cost"
        ]
      }
    },
...
  }
}